+++ deno.lock
@@ -136,6 +136,7 @@
         "jsr:@std/cli",
         "jsr:@std/fs@0.221",
         "jsr:@std/semver",
+        "npm:@octokit/types",
         "npm:octokit"
       ]
     },
