+++ src/tests/commands/init/aws/eks.test.ts
@@ -1,0 +1,35 @@
+import { assert } from "test-deps";
+import { runCndi } from "src/tests/helpers/run-cndi.ts";
+
+Deno.env.set("CNDI_TELEMETRY", "debug");
+
+const ogDir = Deno.cwd();
+
+const cleanup = () => {
+  Deno.chdir(ogDir);
+};
+
+Deno.test(
+  "'cndi init -t basic -l aws/eks should succeed",
+  async (t) => {
+    let dir = "";
+
+    await t.step("setup", async () => {
+      dir = await Deno.makeTempDir();
+      Deno.chdir(dir);
+    });
+
+    await t.step("test", async () => {
+      const { status } = await runCndi(
+        "init",
+        "-t",
+        "basic",
+        "-l",
+        "aws/eks",
+      );
+      assert(status.success);
+    });
+
+    await t.step("cleanup", cleanup);
+  },
+);
